Support for mailchimp3 1.0.12 (for Python 3)
We use the `_put` method in ShoopMailchimpClient.  That is available
through the MailChimp base class with a recent version of mailchimp3
(e.g. 1.0.17), but that version does not support Python 3.  Latest
version of mailchimp3 supporting Python 3 is 1.0.12, but that does not
have the `_put` method.

Provide a implementation for the `_put` method in our derived class so
that our implementation works also with mailchimp3 1.0.12.  Also update
the setup.py to be more specific about required versions.

Refs SHOOP-2570